Enable blob caching in /var/cache/blobs by default
Use /var/cache/blobs, by default, for caching blobs.  Blob caching is
useful for avoiding the work of recompressing layers that we inherit
from a base image when we go to push the newly-built image.
If $BUILD_BLOBCACHE_DIR is set, we its value as the location, instead of
/var/cache/blobs.  Setting the variable to an empty value will disable
the cache.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>